+++ NEWS.md
@@ -1,3 +1,9 @@
+# TeacupCerberus 0.1.1
+ 
+* Added functions tc2iqlrvar and tc2phi which replace the prior use of 
+  RcppCoDA::clrvar2iqlrvar and RcppCoDA::clrvar2phi in Vignette. 
+
+
 # TeacupCerberus 0.1
 
 * Added function `teacup_cerberus`